Clean for v1
